refactor: relocate and update Chakra theme path
Moved Chakra theme files from `chakra-ui/theme` to `lib/chakra-ui/theme`. Updated all related imports and package.json script paths. This change improves project structure and ensures better organization of shared library resources.
